Abstract Form UI to a separate file to reduce complexity inside the panes.
